+++ geomjax/rmhmc/__init__.py
@@ -1,0 +1,3 @@
+from . import rmhmc
+
+__all__ = ["rmhmc"]
